Removed problematic naming of setMap method on layers
For the purpose of internal storing and accessing of the map inside the layer methods named 'provideMap' and 'getProvidedMap' are introduced.
Explanation: The setMap method is used in openlayers to make a layer unmanaged and this is not wanted in this case. The new behaviour reduces management overhead and errors.